+++ src/lib.rs
@@ -48,6 +48,7 @@ pub mod glb;
 pub mod json_model;
 pub mod json_to_scene;
 pub mod scene_to_json;
+pub mod validation;
 
 pub use asset_source::BufferViewAsset;
 pub use decoder::GltfDecoder;